pyproject.toml
[project]
name = "ainews-clustering"
version = "0.1.0"
description = "News dataset exploration with clustering"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "ipython>=8.0.0",
    "scikit-learn>=1.5.2",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "numba>=0.60",
    "llvmlite>=0.43",
    "renumics-spotlight>=1.7.0",
    "openai>=2.15.0",
    "tabulate>=0.9.0",
    "scipy>=1.16.3",
    "torch>=2.9.1",
    "requests>=2.32.5",
    "pgvector>=0.4.2",
    "umap-learn>=0.5.11",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.2.2",
    "ray[client]==2.43.0",
    "numpy<2.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
